latest changes

+++ database/migrations/2023_04_11_201152_add_column_user_id_to_projects_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('projects', function (Blueprint $table) {
-            $table->unsignedBigInteger('user_id')->nullable();
+            $table->integer('user_id')->nullable();
         });
     }
